backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-multipart",
    "fastapi",
    "qdrant-client",
    "uvicorn",
    "pytesseract",
    "requests",
    "pdf2image",
    "pypdf",
    "sentence-transformers",
    "d",
    "langchain-openai",
    "openai",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



